slirpvde was in trouble on 64bit architectures (tnx Andrea Arcangeli for the bugfix).

+++ trunk/vde-2/slirpvde/bootp.h
@@ -98,7 +98,7 @@ struct bootp_t {
     uint8_t bp_htype;
     uint8_t bp_hlen;
     uint8_t bp_hops;
-    unsigned long bp_xid;
+    unsigned int bp_xid;
     unsigned short bp_secs;
     unsigned short unused;
     struct in_addr bp_ciaddr;
